Refactor sidebar.py into browser folder
